refactor: 함수 분리, 함수명 변경
