+++ src/js/app.js
@@ -1,4 +1,4 @@
-import LottoView from './views/LottoView.js';
+import LottoView from './ui/LottoView.js';
 
 window.addEventListener('DOMContentLoaded', () => {
   const lottoView = new LottoView();
