+++ src/js/domains/LottoMachine.js
@@ -48,7 +48,7 @@ export default class LottoMachine {
       .fill()
       .map(() => {
         const lotto = new Lotto();
-        lotto.numbers = lotto.generateNumbers(lotto.generateRandomNumber);
+        lotto.numbers = lotto.pickNumbers(lotto.generateRandomNumber);
         return lotto;
       });
   }
